requirements.txt
# Eagle Eye Sales Intelligence - Dependencies
# Install with: pip install -r requirements.txt

# Core Framework
streamlit>=1.28.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.15.0
folium>=0.14.0
streamlit-folium>=0.13.0

# API Integrations
requests>=2.31.0
googlemaps>=4.10.0

# Environment & Config
python-dotenv>=1.0.0

# Excel Export
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Legacy dependencies (may not be needed)
PyYAML>=6.0.0
gspread>=5.10.0
google-auth>=2.22.0

# Optional Performance Improvements
# Uncomment if needed:
# streamlit-aggrid>=0.3.4
# plotly-resampler>=0.9.0

# Development Tools (Optional)
# black>=23.0.0
# flake8>=6.0.0
# pytest>=7.4.0
